feat: déploiement en production

See merge request pwa-group-b/doremix!92

+++ src/back/controllers/playlist.py
@@ -1,6 +1,7 @@
 from sqlalchemy.orm import Session
 from repositories import PlaylistRepository
 from fastapi import HTTPException, UploadFile, Response
+from models.enums import PlaylistVisibility
 from utils.image_processor import save_cover_image
 from models.playlist import Playlist
 from models import User
@@ -12,26 +13,30 @@ class PlaylistController:
         return PlaylistRepository.get_accessible_playlists(db, user_id)
 
     @staticmethod
-    def get_public_playlists(db: Session):
-        return PlaylistRepository.get_public_playlists(db)
+    def get_public_playlists(db: Session, user: User):
+        return PlaylistRepository.get_public_playlists(db, user)
 
     @staticmethod
-    def get_playlist(db: Session, playlist_id: int, user_id: int):
-        playlist = PlaylistRepository.get_by_id(db, playlist_id, user_id)
+    def get_shared_playlists(db: Session, user_id: int):
+        return PlaylistRepository.get_shared_playlist(db, user_id)
+
+    @staticmethod
+    def get_playlist(db: Session, playlist_id: int, user: User):
+        playlist = PlaylistRepository.get_by_id(db, playlist_id, user)
         if not playlist:
             raise HTTPException(status_code=404, detail="Playlist not found")
         return playlist
 
     @staticmethod
-    def get_playlist_tracks(db: Session, playlist_id: int, user_id: int):
-        playlist = PlaylistRepository.get_by_id(db, playlist_id, user_id)
+    def get_playlist_tracks(db: Session, playlist_id: int, user: User):
+        playlist = PlaylistRepository.get_by_id(db, playlist_id, user)
         if not playlist:
             raise HTTPException(status_code=404, detail="Playlist not found")
         return PlaylistRepository.get_playlist_tracks(db, playlist_id)
 
     @staticmethod
-    def upload_cover(db: Session, playlist_id: int, file: UploadFile, user_id: int):
-        playlist = PlaylistRepository.get_by_id(db, playlist_id, user_id)
+    def upload_cover(db: Session, playlist_id: int, file: UploadFile, user: User):
+        playlist = PlaylistRepository.get_by_id(db, playlist_id, user)
         if not playlist:
             raise HTTPException(status_code=404, detail="Playlist not found")
 
@@ -44,8 +49,8 @@ class PlaylistController:
         return updated_playlist
 
     @staticmethod
-    def remove_track(db: Session, playlist_id: int, track_id: int, user_id: int):
-        playlist = PlaylistRepository.get_by_id(db, playlist_id, user_id)
+    def remove_track(db: Session, playlist_id: int, track_id: int, user: User):
+        playlist = PlaylistRepository.get_by_id(db, playlist_id, user)
         if not playlist:
             raise HTTPException(status_code=404, detail="Playlist not found")
         if not PlaylistRepository.remove_track(db, playlist_id, track_id):
@@ -79,36 +84,43 @@ class PlaylistController:
         return PlaylistRepository.create(db, new_playlist)
 
     @staticmethod
-    def delete_playlist(db: Session, playlist_id: int, user_id: int):
-        # TODO: Quand l'auth sera en place, ajouter user_id en paramètre :
-        # def delete_playlist(db: Session, playlist_id: int, user_id: int):
-
-        playlist = PlaylistRepository.get_by_id(db, playlist_id, user_id)
+    def delete_playlist(db: Session, playlist_id: int, user: User):
+        playlist = PlaylistRepository.get_by_id_raw(db, playlist_id)
 
         if not playlist:
             raise HTTPException(status_code=404, detail="Playlist not found")
 
-        # TODO: Quand l'auth sera en place, vérifier que l'utilisateur est le propriétaire :
-        # if playlist.idOwner != user_id:
-        #     raise HTTPException(status_code=403, detail="You are not the owner of this playlist")
+        # Only playlist owner and admin can delete playlist
+        if not (playlist.idOwner == user.idUser or user.idRole == 3):
+            raise HTTPException(
+                status_code=403, detail="You're not allowed to delete this playlist."
+            )
 
         PlaylistRepository.delete(db, playlist)
 
         return {"message": f"Playlist '{playlist.name}' successfully deleted"}
 
     @staticmethod
-    def update_playlist(db: Session, playlist_id: int, update_data: dict, user_id: int):
-        # TODO: Quand l'auth sera en place, ajouter user_id en paramètre :
-        # def update_playlist(db: Session, playlist_id: int, update_data: dict, user_id: int):
-
-        playlist = PlaylistRepository.get_by_id(db, playlist_id, user_id)
+    def update_playlist(db: Session, playlist_id: int, update_data: dict, user: User):
+        playlist = PlaylistRepository.get_by_id(db, playlist_id, user)
+        users, _ = PlaylistRepository.list_shared_user(db, playlist_id, user.idUser)
+        editors = [user.idUser for user in users if user.editor]
 
         if not playlist:
             raise HTTPException(status_code=404, detail="Playlist not found")
 
-        # TODO: Quand l'auth sera en place, vérifier que l'utilisateur est le propriétaire :
-        # if playlist.idOwner != user_id:
-        #     raise HTTPException(status_code=403, detail="You are not the owner of this playlist")
+        # Only playlist owner, playlist editors and admin can update playlist
+        if not (
+            playlist.idOwner == user.idUser
+            or user.idRole == 3
+            or user.idUser in editors
+        ):
+            raise HTTPException(
+                status_code=404, detail="You're not allowed to update this playlist."
+            )
+
+        if update_data["visibility"] == PlaylistVisibility.OPEN:
+            update_data["idOwner"] = 1
 
         return PlaylistRepository.update_playlist(db, playlist, update_data)
 
@@ -134,6 +146,15 @@ class PlaylistController:
             raise HTTPException(500, "Failed to add track")
 
         return track
+
+    @staticmethod
+    def shared_with(db: Session, playlist_id: int, current_user_id: int):
+        users, err = PlaylistRepository.list_shared_user(
+            db, playlist_id, current_user_id
+        )
+        if err:
+            raise HTTPException(403, err)
+        return users
 
     @staticmethod
     def share_user(
